+++ fluss-lake/fluss-lake-paimon/src/test/java/com/alibaba/fluss/lake/paimon/source/PaimonSplitPlannerTest.java
@@ -20,7 +20,6 @@
 
 package com.alibaba.fluss.lake.paimon.source;
 
-import com.alibaba.fluss.lake.paimon.flink.PaimonLakeHouseTestBase;
 import com.alibaba.fluss.lake.source.LakeSource;
 import com.alibaba.fluss.metadata.TablePath;
 
@@ -43,7 +42,7 @@ import java.util.stream.Collectors;
 import static org.assertj.core.api.Assertions.assertThat;
 
 /** Test case for {@link PaimonSplitPlanner}. */
-class PaimonSplitPlannerTest extends PaimonLakeHouseTestBase {
+class PaimonSplitPlannerTest extends PaimonSourceTestBase {
     @Test
     void testPlan() throws Exception {
         // prepare paimon table
